requirements.txt
fastai==1.0.61
torch==1.6.0
torchvision==0.7.0
opencv-python
pillow
requests
numpy
tqdm